Adding two attributes in street_network section for geovelo
